Restructure output related to actions/rules.
Previously, the output record had keys for actions ('report' and 'noop'). The
corresponding values were lists of rule descriptions that resulted in the
action. Untriggered rules weren't mentioned.

This commit changes the output record to have fields for each rule. The
corresponding value is the outcome of the rule: either the action ('report',
'noop'), or the special 'rule-not-triggered' value indicating that the rule
didn't trigger.

This makes the output more regular and easier to deal with down the line (no
extra unnesting/processing of the 'report' field, no dealing with null values).